+++ src/main/java/com/app/shahbaztrades/model/entity/StrategyOrder.java
@@ -9,6 +9,7 @@ import lombok.experimental.FieldNameConstants;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.MongoId;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.format.DateTimeFormatter;
 
@@ -30,7 +31,7 @@ public class StrategyOrder {
 
     String strategyName;
 
-    float amount;
+    BigDecimal amount;
 
     BrokerType broker;
